+++ plugin/src/main/java/net/dzikoysk/funnyguilds/feature/placeholders/resolver/GuildResolver.java
@@ -1,0 +1,8 @@
+package net.dzikoysk.funnyguilds.feature.placeholders.resolver;
+
+import net.dzikoysk.funnyguilds.guild.Guild;
+
+@FunctionalInterface
+public interface GuildResolver {
+    Object resolve(Guild guild);
+}
