+++ plugin/src/main/java/net/dzikoysk/funnyguilds/feature/placeholders/resolver/TimeResolver.java
@@ -1,0 +1,8 @@
+package net.dzikoysk.funnyguilds.feature.placeholders.resolver;
+
+import java.time.LocalDateTime;
+
+@FunctionalInterface
+public interface TimeResolver {
+    Object resolve(LocalDateTime now);
+}
