+++ plugin/src/main/java/net/dzikoysk/funnyguilds/feature/placeholders/resolver/UserResolver.java
@@ -1,0 +1,8 @@
+package net.dzikoysk.funnyguilds.feature.placeholders.resolver;
+
+import net.dzikoysk.funnyguilds.user.User;
+
+@FunctionalInterface
+public interface UserResolver {
+    Object resolve(User user);
+}
