+++ plugin/src/main/java/net/dzikoysk/funnyguilds/feature/prefix/IndividualPrefix.java
@@ -3,7 +3,7 @@ package net.dzikoysk.funnyguilds.feature.prefix;
 import java.util.Set;
 import net.dzikoysk.funnyguilds.FunnyGuilds;
 import net.dzikoysk.funnyguilds.config.PluginConfiguration;
-import net.dzikoysk.funnyguilds.feature.placeholders.Placeholders;
+import net.dzikoysk.funnyguilds.feature.placeholders.legacy.Placeholders;
 import net.dzikoysk.funnyguilds.guild.Guild;
 import net.dzikoysk.funnyguilds.user.User;
 import org.bukkit.scoreboard.Scoreboard;
